introduce mutable handle info
