+++ client-spark/spark2/src/main/java/org/apache/spark/shuffle/RssShuffleManager.java
@@ -795,7 +795,7 @@ public class RssShuffleManager extends RssShuffleManagerBase {
   }
 
   @Override
-  public ShuffleHandleInfo getShuffleHandleInfoByShuffleId(int shuffleId) {
+  public ShuffleHandleInfoBase getShuffleHandleInfoByShuffleId(int shuffleId) {
     return shuffleIdToShuffleHandleInfo.get(shuffleId);
   }
